Improved structure, use modules
